Move functions / classes edge -> web
UIA.Web is a new base for browser specific implementations:
- spartan-edge
- Chromium based browsers: anaheim-edge, chrome

This commit is a direct move (no alterations) of code to preserve history.

- Move function splitUIAElementAttribs
- Move EdgeTextInfo
- Move shareable aspects of class EdgeNode:
  * _get_role
  * _get_states
  * _get_ariaProperties
  * _get_isCurrent
  * _get_roleText
  * _get_placeholder
  * _get_landmark
- Move class EdgeList
- Move class EdgeHeadingQuickNavItem
- Move class EdgeHeadingQuicknavIterator
- Move shareable aspects of class EdgeHTMLTreeInterceptor:
  * makeTextInfo
  * shouldPassThrough
  * _iterNodesByType
